examples/flet_image_gallery_saver_example/pyproject.toml
[project]
name = "flet-image-gallery-saver-example"
version = "1.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Flet developer", email = "you@example.com" }
]
dependencies = [
  "flet-image-gallery-saver",
  "flet>=0.28.2",
  "pillow>=11.2.1",
]

[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "com.mycompany"

# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "flet-image-gallery-saver-example"

# company name to display in about app dialogs
company = "Flet"

# copyright text to display in about app dialogs
copyright = "Copyright (C) 2025 by MyCompany"

[tool.flet.app]
path = "src"

[tool.flet.dev_packages]
flet-image-gallery-saver = "..\\.." # relative path

[tool.uv]
dev-dependencies = [
    "flet[all]==0.28.2",
]

[tool.uv.sources]
flet-image-gallery-saver = { path = "..\\..\\", editable = true }

[tool.poetry.group.dev.dependencies]
flet = {extras = ["all"], version = "0.28.2"}
flet-image-gallery-saver = {path = "..\\..\\", develop = true}
